Refactor according to es6 instruction
