+++ index.html
@@ -5,7 +5,9 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="index.css" />
-    <title>Microverse</title>
+    <title>Awesome Book with ES6</title>
+
+    <script src="index.js" defer type="module"></script>
   </head>
   <body class="show-list">
     <header class="border-box">
@@ -56,7 +58,5 @@
       Copyright 2023: Lorem ipsum dolor sit, amet consectetur adipisicing elit.
       Vitae doloremque veniam laboriosam? Adipisci, velit asperiores.
     </footer>
-
-    <script src="script.js"></script>
   </body>
 </html>
