+++ modules/AwesomeBook.js
@@ -63,43 +63,4 @@ class AwesomeBook {
   };
 }
 
-const form = document.querySelector('form');
-const awesomeBook = new AwesomeBook();
-
-form.addEventListener('submit', (e) => {
-  e.preventDefault();
-  const [titleInput, authorInput] = e.target.elements;
-  awesomeBook.insert({
-    author: authorInput.value,
-    title: titleInput.value,
-  });
-  e.target.reset();
-});
-
-// handle nav buttons
-const navBtns = document.querySelectorAll('.nav-button');
-
-navBtns.forEach((btn) => {
-  btn.addEventListener('click', (e) => {
-    const text = e.currentTarget.innerText;
-    document.body.className = `show-${text.toLowerCase().replace(' ', '-')}`;
-  });
-});
-
-// Update time element every second
-const options = {
-  year: 'numeric',
-  month: 'long',
-  weekday: 'long',
-  day: 'numeric',
-  hour: 'numeric',
-  minute: 'numeric',
-  second: 'numeric',
-  hour12: true,
-};
-
-const timeEl = document.querySelector('time');
-
-setInterval(() => {
-  timeEl.innerText = new Date().toLocaleDateString('en-US', options);
-}, 1000);
+export default AwesomeBook;
